backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=2.0.0
httpx>=0.28.0
reportlab>=4.2.0
python-docx>=1.1.0
python-dotenv>=1.0.1
pydantic>=2.7.0
